Move all 'declare global' to index.ts
This is really important for libraries - if TypeScript never imports a
file (because it doesn't need to, for whatever reason) then the types
will not be imported, and thus global declaration merging won't work for
consumers. It's strongly recommended that library authors put their
`declare global` in the `index.ts` file.